feat(nav): lean primary nav; move Data under Admin; drop Tasks/Maturity/Decisions entries
Product call (2026-07-10): trim the left nav to Home / Codebases / Learning / Query / Settings.
- Tasks, Maturity, Decisions removed from the nav (routes still resolve by direct URL; Decisions
  is empty + unused, full backend teardown tracked in docs/TODO.md).
- Data (the ingested-data channel browser) moves from the primary nav to an Admin → Data tab, so
  it's admin-gated — it's a verification/debug view, not a daily surface. The browser logic is
  extracted into a shared `DataBrowser` server component; `/library` index redirects to
  `/admin/data` (preserving ?channel/?limit); `/library/[id]` detail + `/library/skills` stay put.
- The now-empty "Work" nav group is dropped.

Also records the full navigation + Learning-page plan in docs/TODO.md.
